The move script now correctly moves the codeigniter files into the desired location

+++ utils/move.py
@@ -4,5 +4,9 @@ import shutil
 source = '../CodeIgniter-3.1.8/'
 destination = '../'
 
+# Move the contents of the codeigniter folder to another location
 for file_name in os.listdir(source):
-    shutil.move(source + file_name, destination)
+    if not os.path.exists(destination + file_name):
+        shutil.move(source + file_name, destination)
+
+shutil.rmtree(source)
